initial header webcomponent impl
